server/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
mongoengine==0.27.0
djongo==1.3.6
pymongo==4.6.0
python-decouple==3.8
django-environ==0.11.2
gunicorn==21.2.0
python-dotenv==1.0.0
